Extract omop helper to db
